% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_support_object.R
\name{as.data.frame.rna_support}
\alias{as.data.frame.rna_support}
\title{Convert rna_support Object to Data Frame}
\usage{
\method{as.data.frame}{rna_support}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rna_support}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A data frame with one row and columns corresponding to the components
of the \code{rna_support} object.
}
\description{
This function converts an object of class \code{rna_support} to a data frame.
}
