% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_support.R
\name{compute_rna_support_ratio}
\alias{compute_rna_support_ratio}
\title{Compute proportion of DNA mutations with RNA support}
\usage{
compute_rna_support_ratio(
  data,
  col_DNA_AF = "DNA_AF",
  col_RNA_DP = "RNA_DP",
  col_RNA_AD = "RNA_AD",
  min_alt_supporting_reads = 1,
  confidence = 0.95,
  simple = FALSE
)
}
\arguments{
\item{data}{A dataframe representing all autosomal mutations present in 1 sample.
Each row should represent a different mutation. Must the following columns (order does NOT matter, just column names)
\enumerate{
\item \strong{col_DNA_AF} Raw allele frequency in DNA (not adjusted for purity/ploidy). AF in oncoanalyser.
\item \strong{col_RNA_DP} Total depth of coverage at loci in RNA. DP in oncoanalyser.
\item \strong{col_RNA_AD} Number of reads supporting ALT allele in RNA. AD in oncoanalyser.
}}

\item{col_DNA_AF}{name of DNA_AF column (string)}

\item{col_RNA_DP}{name of RNA_DP column (string)}

\item{col_RNA_AD}{name of RNA_AD column (string)}

\item{min_alt_supporting_reads}{the minimum number of alt supporting reads
required to consider a DNA variant supported by RNA (numeric)}

\item{confidence}{For any particular variant, the theoretical percentage of
finding >=\code{min_alt_supporting_reads} above which we are comfortable
classifying the variant as 'expected' to have RNA support.}

\item{simple}{just return Psupp instead of a full rna_support obj}
}
\value{
A list of class \code{rna_support} containing the following components:
\describe{
\item{\strong{total_variants_evaluated}}{Integer. The total number of variants evaluated for RNA support.}
\item{\strong{expected_rna_support_variants}}{Integer. The number of variants expected to show RNA support based on their DNA characteristics (e.g., variant allele frequency, read depth).}
\item{\strong{proportion_with_observed_support}}{Numeric. The proportion of variants expected to show RNA support where RNA support was actually observed. Sometimes called 'Psupp'}
\item{\strong{proportion_unexpected_with_observed_support}}{Numeric. The proportion of variants not expected to show RNA support that nonetheless had RNA support observed.}
}
}
\description{
In what proportion of the DNA mutations which theoretically should have RNA support
do we observe that support.
Expectation of RNA support is based on DNA allele frequency and RNA depth.
Observed RNA support is based on presense of >= \code{min_alt_supporting_reads}.
}
\examples{
# Create a sample data frame representing mutations in a single sample
sample_data <- data.frame(
  DNA_AF = c(0.5, 0.3, 0.2, 0.1),  # DNA variant allele frequencies
  RNA_DP = c(100, 50, 20, 10),     # RNA read depths at each locus
  RNA_AD = c(48, 15, 4, 0)         # RNA reads supporting the alternate allele
)

# Calculate Psupp for the sample
Psupp <- compute_rna_support_ratio(
  data = sample_data,
  col_DNA_AF = "DNA_AF",
  col_RNA_DP = "RNA_DP",
  col_RNA_AD = "RNA_AD",
  min_alt_supporting_reads = 2,
  confidence = 0.95
)
}
