% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_support_object.R
\name{create_rna_support_result}
\alias{create_rna_support_result}
\title{Create RNA Support Result}
\usage{
create_rna_support_result(
  total_variants,
  expected_variants,
  observed_support_proportion,
  unexpected_support_proportion
)
}
\arguments{
\item{total_variants}{Integer. The total number of variants evaluated.}

\item{expected_variants}{Integer. The number of variants expected to have RNA support.}

\item{observed_support_proportion}{Numeric. The proportion of expected variants with RNA support observed.}

\item{unexpected_support_proportion}{Numeric. The proportion of unexpected variants that still had RNA support.}
}
\value{
A list of class \code{rna_support} containing the following components:
\describe{
\item{\strong{total_variants_evaluated}}{Integer. The total number of variants evaluated for RNA support.}
\item{\strong{expected_rna_support_variants}}{Integer. The number of variants expected to show RNA support based on their DNA characteristics (e.g., variant allele frequency, read depth).}
\item{\strong{proportion_with_observed_support}}{Numeric. The proportion of variants expected to show RNA support where RNA support was actually observed. Sometimes called 'Psupp'}
\item{\strong{proportion_unexpected_with_observed_support}}{Numeric. The proportion of variants not expected to show RNA support that nonetheless had RNA support observed.}
}
}
\description{
This function creates an object of class \code{rna_support_result} which holds
the results of the RNA support evaluation, including the total number of
variants evaluated, expected RNA support variants, the proportion of variants
with observed support (expected & unexpected).
}
\examples{
create_rna_support_result(
  total_variants = 100,
  expected_variants = 80,
  observed_support_proportion = 0.100,
  unexpected_support_proportion = 0.10
)
}
