% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_support_object.R
\name{print.rna_support}
\alias{print.rna_support}
\title{Print method for rna_support object}
\usage{
\method{print}{rna_support}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rna_support}.}

\item{...}{Additional arguments passed to or from other methods (not used here).}
}
\description{
Custom print method to display the properties of an rna_support object.
}
\keyword{internal}
