% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{sex_chromosomes}
\alias{sex_chromosomes}
\title{Common Sex Chromosome Names}
\usage{
sex_chromosomes()
}
\value{
a vector of common sex chromosome names
}
\description{
Common Sex Chromosome Names
}
\examples{
sex_chromosomes()
}
