% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{write_simulated_vcfs}
\alias{write_simulated_vcfs}
\title{Generates a bunch simulated VCFs}
\usage{
write_simulated_vcfs(
  outfolder = "cohort/",
  prefix = "simulated",
  nsamples = 20,
  nvariants = 20,
  seed = 111
)
}
\arguments{
\item{outfolder}{folder to write simulated VCFs to}

\item{prefix}{VCF name prefixes}

\item{nsamples}{how many VCFs to make}

\item{nvariants}{how many variants to write to each VCF}
}
\value{
invisibly returns NULL. run for its side effects
}
\description{
Used to test running in cohort mode from manifest
}
\examples{
if(interactive){
write_simulated_vcfs(outfolder="cohort")
}
}
