% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_support.R
\name{compute_rna_support_ratio_from_manifest}
\alias{compute_rna_support_ratio_from_manifest}
\title{Compute RNA support from multiple VCFs}
\usage{
compute_rna_support_ratio_from_manifest(
  manifest,
  min_alt_supporting_reads = 2,
  confidence = 0.95,
  field_AF = "AF",
  field_DP = "DP",
  field_AD = "AD",
  pass_only = TRUE,
  sep = "\\t",
  exclude_sex_chromosomes = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{manifest}{path to tsv with the following columns
\enumerate{
\item vcf: path to a multi sample VCF which describes mutations in both DNA and RNA.
These are typically produced by calling the \href{https://github.com/hartwigmedical/hmftools/blob/master/sage/README.md#append-mode}{sage variant caller in append mode}.
\item dna_sample: the name of the tumour DNA sample in the VCF
\item rna_sample: the name of the tumour RNA sample in the VCF
}}

\item{min_alt_supporting_reads}{the minimum number of alt supporting reads
required to consider a DNA variant supported by RNA (numeric)}

\item{confidence}{For any particular variant, the theoretical percentage of
finding >=\code{min_alt_supporting_reads} above which we are comfortable
classifying the variant as 'expected' to have RNA support.}

\item{field_AF}{Genotype field that describes allele frequency uncorrected for tumour purity/ploidy. (string)}

\item{field_DP}{Genotype field that describes read depth. (string)}

\item{field_AD}{Genotype field that describes allelic depth. (string)}

\item{pass_only}{only compute rna support ratio using variants where FILTER=PASS in VCF (flag)}

\item{sep}{delimiter used in manifest file.}

\item{exclude_sex_chromosomes}{Exclude variants from sex chromosomes. See \code{\link[=sex_chromosomes]{sex_chromosomes()}} for all values of 'CHROM' we would identify as sex chromosomes. (flag)}

\item{verbose}{verbose (flag)}
}
\value{
the manifest dataframe with an extra column 'rna_support_ratio' describing
\strong{Psupp}: The proportion of variants in the sample where we expected to
see  RNA evidence for and observed that RNA evidence in practice.
}
\description{
In what proportion of the DNA mutations which theoretically should have RNA support
do we observe that support.
Expectation of RNA support is based on DNA allele frequency and RNA depth.
Observed RNA support is based on presense of >= \code{min_alt_supporting_reads}.
}
\examples{
if(interactive()){
  manifest <- system.file("manifest.tsv", package = "Rtifact")
  write_simulated_vcfs("cohort")
  compute_rna_support_ratio_from_manifest(
    manifest,
    field_AF = "AF",
    field_DP = "DP",
    field_AD = "AD",
    min_alt_supporting_reads = 2,
    confidence = 0.95
  )
}

}
