% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_support.R
\name{compute_rna_support_ratio_from_vcf}
\alias{compute_rna_support_ratio_from_vcf}
\title{Compute proportion of DNA mutations with RNA support}
\usage{
compute_rna_support_ratio_from_vcf(
  vcf,
  dna_sample,
  rna_sample,
  min_alt_supporting_reads = 2,
  confidence = 0.95,
  field_AF = "AF",
  field_DP = "DP",
  field_AD = "AD",
  pass_only = TRUE,
  return_data = FALSE,
  verbose = TRUE,
  exclude_sex_chromosomes = TRUE
)
}
\arguments{
\item{vcf}{path to a multi sample VCF which describes mutations in both DNA and RNA samples.
These can be produced by running the \href{https://github.com/hartwigmedical/hmftools/blob/master/sage/README.md#append-mode}{sage variant caller in append mode}.}

\item{dna_sample}{the name of the tumour DNA sample in the VCF}

\item{rna_sample}{the name of the tumour RNA sample in the VCF}

\item{min_alt_supporting_reads}{the minimum number of alt supporting reads
required to consider a DNA variant supported by RNA (numeric)}

\item{confidence}{For any particular variant, the theoretical percentage of
finding >=\code{min_alt_supporting_reads} above which we are comfortable
classifying the variant as 'expected' to have RNA support.}

\item{field_AF}{Genotype field that describes allele frequency uncorrected for tumour purity/ploidy. (string)}

\item{field_DP}{Genotype field that describes read depth. (string)}

\item{field_AD}{Genotype field that describes allelic depth. (string)}

\item{pass_only}{only compute rna support ratio using variants where FILTER=PASS in VCF (flag)}

\item{return_data}{return the data.frame supplied to \code{\link[=compute_rna_support_ratio]{compute_rna_support_ratio()}} instead of the computed result. Mainly used for debug (flag)}

\item{verbose}{verbose (flag)}

\item{exclude_sex_chromosomes}{Exclude variants from sex chromosomes. See \code{\link[=sex_chromosomes]{sex_chromosomes()}} for all values of 'CHROM' we would identify as sex chromosomes. (flag)}
}
\description{
Compute proportion of DNA mutations with RNA support
}
\examples{
vcf <- system.file("example.vcf", package = "Rtifact")
compute_rna_support_ratio_from_vcf(vcf, dna_sample = "DNA", rna_sample = "RNA")
}
