% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_support.R
\name{probability_of_rna_support}
\alias{probability_of_rna_support}
\title{Probability a DNA mutation will have RNA support}
\usage{
probability_of_rna_support(DNA_AF, RNA_DP, min_alt_supporting_reads = 2)
}
\arguments{
\item{DNA_AF}{the unadjusted allele frequency of the variant (numeric)}

\item{RNA_DP}{the total number of RNA reads covering the Loci (numeric)}

\item{min_alt_supporting_reads}{the minimum number of alt supporting reads (numeric)}
}
\value{
a vector of probabilities (proportions)
}
\description{
Probability a DNA mutation will have RNA support
}
\examples{
# Single Variant
probability_of_rna_support(DNA_AF = 0.5, RNA_DP = 1, min_alt_supporting_reads = 1)

# Multiple Variants
probability_of_rna_support(
  DNA_AF = c(0.5, 0.5, 0.1),
  RNA_DP = c(1, 2, 2),
  min_alt_supporting_reads = c(1, 1, 2)
)

# Multiple variants using the same setting for 'min_alt_supporting_reads'
probability_of_rna_support(
  DNA_AF = c(0.5, 0.5),
  RNA_DP = c(1, 2),
  min_alt_supporting_reads = 1
)
}
